requirements.txt
discord.py
pytube
aiohttp
Pillow
llama-cpp-python
python-dotenv 